bunch of 64-bit compile fixes

+++ antzplus/ANTzBridge/antzcsvwriter.cpp
@@ -422,7 +422,7 @@ namespace SynGlyphXANTz {
 			file.WriteLine(grid);
 		}
 
-		return firstId + grids.size();
+		return firstId + (unsigned long)grids.size();
 	}
 
 	ANTzCSVWriter& ANTzCSVWriter::GetInstance() {
